Moved spritesheet data to zero dependancy script (NEED TO UPDATE ALL REFS)
